+++ ops/logging.go
@@ -1,6 +1,7 @@
 package ops
 
 import (
+	"bytes"
 	"context"
 
 	"github.com/valyala/fasthttp"
@@ -21,19 +22,16 @@ func AddLoggingContext(c userContextProvider, key, value string) {
 	})
 }
 
-// LogHTTPEvent returned function adds request and response data to the zerolog Event
-func LogHTTPEvent(req *fasthttp.Request, resp *fasthttp.Response, err error) func(event *zerolog.Event) {
+// LogHTTPRequest return a function that adds request data to the zerolog Event
+func LogHTTPRequest(req *fasthttp.Request) func(event *zerolog.Event) {
 	return func(event *zerolog.Event) {
-		event.Err(err)
-
 		// Google cloud structured logging recognize the following fields:
 		// https://cloud.google.com/logging/docs/structured-logging#special-payload-fields
-		requestEvent := zerolog.Dict()
+		requestDict := zerolog.Dict()
 
 		// Standard http request fields
-		requestEvent.Bytes("requestUrl", req.URI().FullURI()).
+		requestDict.Bytes("requestUrl", req.URI().FullURI()).
 			Bytes("requestMethod", req.Header.Method()).
-			Int("status", resp.StatusCode()).
 			Bytes("protocol", req.Header.Protocol())
 
 		// Request headers
@@ -44,18 +42,38 @@ func LogHTTPEvent(req *fasthttp.Request, resp *fasthttp.Response, err error) fun
 		if contentEncoding := req.Header.ContentEncoding(); contentEncoding != nil {
 			requestHeaders.Bytes("Content-Encoding", contentEncoding)
 		}
-		requestEvent.Dict("requestHeaders", requestHeaders)
+		requestDict.Dict("requestHeaders", requestHeaders)
 
 		// Request body
 		if body := req.Body(); body != nil {
 			// content is encoded, don't bother decompressing
 			if encoding := req.Header.ContentEncoding(); encoding != nil {
-				requestEvent.Bytes("request", encoding)
+				requestDict.Bytes("request", encoding)
 			} else if contentType := req.Header.ContentType(); isContentTypeLogged(contentType) {
-				requestEvent.RawJSON("request", body)
+				if isContentTypeJSON(contentType) {
+					requestDict.RawJSON("request", body)
+				} else {
+					requestDict.Bytes("request", body)
+				}
 			}
-			requestEvent.Int("requestSize", req.Header.ContentLength())
+			requestDict.Int("requestSize", req.Header.ContentLength())
 		}
+
+		event.Dict("httpRequest", requestDict)
+	}
+}
+
+// LogHTTPResponse return a function that adds response data to the zerolog Event
+func LogHTTPResponse(resp *fasthttp.Response, err error) func(event *zerolog.Event) {
+	return func(event *zerolog.Event) {
+		event.Err(err)
+
+		// Google cloud structured logging recognize the following fields:
+		// https://cloud.google.com/logging/docs/structured-logging#special-payload-fields
+		responseDict := zerolog.Dict()
+
+		// Standard http response fields
+		responseDict.Int("status", resp.StatusCode())
 
 		// Response headers
 		responseHeaders := zerolog.Dict()
@@ -65,19 +83,23 @@ func LogHTTPEvent(req *fasthttp.Request, resp *fasthttp.Response, err error) fun
 		if contentEncoding := resp.Header.ContentEncoding(); contentEncoding != nil {
 			responseHeaders.Bytes("Content-Encoding", contentEncoding)
 		}
-		requestEvent.Dict("responseHeaders", responseHeaders)
+		responseDict.Dict("responseHeaders", responseHeaders)
 
 		// Response body
 		if body := resp.Body(); body != nil {
 			// content is encoded, don't bother decompressing
 			if encoding := resp.Header.ContentEncoding(); encoding != nil {
-				requestEvent.Bytes("response", encoding)
+				responseDict.Bytes("response", encoding)
 			} else if contentType := resp.Header.ContentType(); isContentTypeLogged(contentType) {
-				requestEvent.RawJSON("response", body)
+				if isContentTypeJSON(contentType) {
+					responseDict.RawJSON("response", body)
+				} else {
+					responseDict.Bytes("response", body)
+				}
 			}
-			requestEvent.Int("responseSize", resp.Header.ContentLength())
+			responseDict.Int("responseSize", resp.Header.ContentLength())
 		}
-		event.Dict("httpRequest", requestEvent)
+		event.Dict("httpResponse", responseDict)
 	}
 }
 
@@ -96,4 +118,9 @@ var loggedContentTypes = map[string]struct{}{
 func isContentTypeLogged(contentType []byte) bool {
 	_, found := loggedContentTypes[string(contentType)]
 	return found
+}
+
+// isContentTypeJSON returns true for content types that contains json data, otherwise false.
+func isContentTypeJSON(contentType []byte) bool {
+	return bytes.Contains(contentType, []byte("/json")) || bytes.Contains(contentType, []byte("+json"))
 }